Update build to use Fabric8 Docker Maven Plugin
This commit replaces the Spotify Docker Maven Plugin with the Fabric8
Plugin. The Fabric8 Docker Maven Plugin supports the use of Docker BuildX
to generate docker images for multiple architectures.

https://dmp.fabric8.io/#build-buildx

This has been configured to support amd64 and arm64 images.

Also note that the Spotify Docker Maven Plugin was archived in November 2022.